projects/flex_cc/pyproject.toml
[tool.poetry]
name = "flex-cc"
version = "0.32.0"
description = ""
authors = ["ComboCurve <support@combocurve.com>"]
classifiers = [
    "Private :: Do not Upload"
]

# tasks in poethepoet package are similar to npm scripts. poetry run poe <taskname>

[tool.poetry.group.dev.dependencies]
pytest-mock-resources = "^2.6.11"
python-on-whales = "^0.60.1"

[tool.poe.tasks]
dev = {cmd = "python -m debugpy --listen 0.0.0.0:5678 -m api.server --wait-for-client --multiprocess -m flask run -h 0.0.0.0", envfile = ".env"}
format = "yapf --diff"
jupyter = "jupyter notebook"
lint = "flake8"
start = "gunicorn -b :$PORT api.server:app"
test = "pytest"
test-unit = 'pytest -m unittest --disable-warnings'
test-integ = 'pytest -m integtest'
test-bench = 'pytest -m benchmark --benchmark-json benchmark-output.json'

[tool.poetry.dependencies]
python = "~3.9"
Flask-Cors = "~3.0"
directaccess = "^1.5.2"
dnspython = "~1.16"
Flask = "^2.0"
google-cloud-bigquery = "^3.2.0"
google-cloud-logging = "^3.1.2"
google-cloud-tasks = "^2.9.1"
google-api-core = "^2.8.0"
google-auth = "^2.6.6"
google-cloud-core = "^2.3.0"
google-cloud-storage = "^2.3.0"
google-resumable-media = "^2.3.3"
googleapis-common-protos = "^1.56.1"
gevent = "^21.12"
gunicorn = {version = "^20.0", extras = ["gevent"]}
greenlet = "^1.1"
mongoengine = "*"
openpyxl = "*"
pandas = "1.3.5"
pusher = "*"
pymongo = "^3.11"
requests = "*"
xlrd = "^1.2"
python-dateutil = "*"
jellyfish = "*"
babel = "*"
pygeohash = "*"
fpdf = "*"
retry = "*"
pyarrow = "^5.0"
jaydebeapi = "^1.2"
jpype1 = "^1.3"
pyexcelerate = "^0.10"
joblib = "~0.17"
scipy = "1.10.0"
redis = "*"
hiredis = "*"
pyshp = "^2.1.3"
pyproj = "^3.3.1"
poethepoet = "^0.12.3"
Shapely = "^1.8.1"
numpy = "1.24.2"
db-dtypes = "==1.0.2"
mapbox = "^0.18.1"
pydantic = "^1.10.2"
mongomock = "^4.1"
polars = "^0.15.18"
grpc-interceptor = "^0.15.1"
jsonschema = "^4.17.3"
google-cloud-secret-manager = "^2.16.2"
launchdarkly-server-sdk = "^8.1.4"

[tool.poetry.dev-dependencies]
flake8 = "5.0.4"
flake8-print = "^4.0.0"
pre-commit = "1.16.1"
yapf = "0.31"
pep8-naming = "^0.11"
python-dotenv = {extras = ["cli"], version = "^0.19.2"}
pytest = "^7.1.2"
pytest-flask = "^1.2.0"
deepdiff = "^5.8.1"
pytest-benchmark = "4.0.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
ortools = "^9.6.2534"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
